refactor(kernel): make phase-7 a shared coordinator step on one deadline
Three repairs to the audited-local boundary, from the cumulative review and
the ownership rule the plan states for phase 7.

**Phase 7 belongs to the shared boundary, not to doctor.** `DoctorPreflight`
took `DoctorPlan` rows and called `pending/1` and `settle/3`, which put
occurrence traversal on the doctor side and let a caller narrow the work by
handing over a shortened list. `LocalPreflight` replaces it: it derives
applicable occurrences itself from descriptors declaring `:audited_local`,
never mentions doctor, and returns `:ok | {:error, diagnostic}` so run and
check can use the same step. Doctor now only settles its rows afterwards,
through the new `DoctorPlan.settle_pending/1`. This also removes the reported
shortcut where an emptied pending list skipped a required check.

**One anchored deadline, never renewed.** Each occurrence previously received
the full timeout, so phase-7 wall time multiplied by the number of occurrences.
The caller now anchors one `Deadline` and every occurrence spends what remains
of it, which is the plan's rule that nested work may narrow a deadline but
never resets it. An exhausted budget fails closed; the catalog has no phase-7
timeout code, so that is reported as an internal error and a dedicated code is
left to the acceptance slice.

**The trio must belong together.** The step validated its prepared run,
catalog, and services independently, so a preparation from one catalog could be
checked against another's declarations with a third's services. It now enforces
the catalog attestation and the runtime-services binding, matching
`ProviderExecution`.

Separately, `Memory.close/1` confirms process identity before sending `:close`
rather than only before terminating: `:close` is a plausible message for an
unrelated GenServer, so a stale handle could previously stop or mutate one. The
foreign-handle test now uses a real GenServer that would honour `:close` and
proves its state is untouched, since a process that ignores calls cannot show
whether anything was sent.

Verified with `mix precommit` and `MIX_ENV=dev mix docs --warnings-as-errors`.
Five invariants mutation-checked: catalog binding, services binding,
applicability filtering, deadline gating, and the store identity guard.
